Merge performed from GSI code version

+++ mocadi/LoopMocadi.pl
@@ -1,48 +1,113 @@
 #! /usr/bin/perl
 
-use lib "/u/crappold/frs/epax/usr/local/share/perl";
-use IO::CaptureOutput qw/capture/;
+
+
+# Author : Christophe Rappold, Dr < c.rappold [()] gsi.de >
+# Copyright 2013 Christophe Rappold
+# License : GPLv3
+
+
+# The script works without any internal modifications. Everything are passed by command line.
+# Example : ./LoopMocadi.pl header_file.in template_file.in mocadi_exe
+# In this way the script can used for any version of mocadi that is available in the system. The input files can adjusted also externally.
+
+# Optional : Please install the corresponding package of perl : 
+# Optional : IO::CaptureOutput from CPAN or your linux distribution package manager (ex: debian package is libio-captureoutput-perl)
+
+use strict;
+use warnings;
+
+#use lib "/u/crappold/frs/epax/usr/local/share/perl";
+#use IO::CaptureOutput qw/capture/;
+#use lib "./usr/local/share/perl";
+use lib "./usr/lib/perl5";
+my $noCapture = 0;
+eval
+{
+# If IO::CaptureOutput is installed 
+    require IO::CaptureOutput;
+    IO::CaptureOutput->import(qw/capture_exec/);
+};
+if($@)
+{
+    $noCapture = 1;
+    die " No IO::CaptureOutput \n";
+    
+}
+else
+{
+    $noCapture = 0;
+    print " IO::CaptureOutput Loaded \n";
+}
+
 my ($stdout, $stderr);
 
+die "Usage: $0 HEADER_FILE [TEMPLATE_FILE] [MOCADI_EXE] [MEANENERGY]\n" if($#ARGV != 3 and $#ARGV != 2 and $#ARGV != 1 and $#ARGV != 0) ;
+
+my $template_mocadi = "template_heb_3.in";
+if($#ARGV == 1 or $#ARGV == 2 or $#ARGV ==3)
+{
+    $template_mocadi = $ARGV[1];
+}
+
+die "TEMPLATEFILE not readable !" if(!-e $template_mocadi);
+die "HEADERFILE not readable !" if(!-e $ARGV[0]);
+die "MOCADI_EXE not executable !" if ($#ARGV == 2 and !-X $ARGV[2]);
+
+my $MeanEnergy =0;
+if($#ARGV == 3)
+{
+    $MeanEnergy = $ARGV[3];
+}
+
+my $mocadi_exe = "./source3.5/mocadi-35";
+if($#ARGV == 2)
+{
+    $mocadi_exe = $ARGV[2];
+}
 
 (my $file_in = $ARGV[0]) =~ s/\.[^.]+$//;
 
-open my $fileheader, '<', $ARGV[0] or die "error opening $filename: $!";
+## -> Load the header file in the variable $header for later concatenation
+
+open my $fileheader, '<', $ARGV[0] or die "error opening $ARGV[0]: $! \n";
 my $header = do { local $/ = undef; <$fileheader> };
 
 close $fileheader;
 
-open my $fileheader, '<', $ARGV[0] or die "error opening $filename: $!";
+## -> Parse the header file to set the fragment of interest $Af $Zf
+
+open my $fileheader, '<', $ARGV[0] or die "error opening $ARGV[0]: $! \n";
 my $Af =0;
 my $Zf =0;
 while(my $lineheader=<$fileheader>)
 {
-    if($lineheader =~ /^.*?([0-9]*)\,.*?([0-9]*)\,.*?Sollfragment$/)
-    {
+    #print $lineheader;
+
+    if($lineheader =~ /^\s*([0-9]*\.[0-9]*|[0-9]*)\,\s*([0-9]*)\,\s*Sollfragment/)
+    {	
+	#print $lineheader;
 	$Af = $1;
 	$Zf = $2;
     }
 }
+if($Af==0 or $Zf==0)
+{
+    die "No valid fragment A=$Af Z=$Zf !\n";
+}
 close $fileheader;
 print "fragment :A".$Af."Z".$Zf."\n";
 
+## -> 1) Open the second file of the mocadi setup : Matrices of the magnets and SAVE statements at the wanted place
+## -> 2) Look for the save statements and save the position for the loop of the SAVE points
 
-# capture sub {
-#     system "./source3.5/mocadi-35 $file_in"; 
-# } => \$stdout, \$stderr;
+open IN2, "<", $template_mocadi or die " error opening $template_mocadi: $! \n";
 
-
-#open IN, $ARGV[0] or die "impossible d'ouvrir !";
-
-open IN2, "<", "template_heb.in" or die " impossible d'ouvrir !";
-
-#open OUT, ">test.in" or die " impossible d'ouvrir !";
-
-@array_save;
+my @array_save;
 while(my $line2=<IN2>)
 {
     my $savestate ;
-    if(($savestate) = ($line2=~ /^ SAVE save(.*?)$/))
+    if(($savestate) = ($line2=~ /^\s*SAVE save(.*?)\s?$/))
     {
 	push(@array_save,$savestate); 
     }
@@ -53,21 +118,35 @@ close IN2;
 print @array_save;
 print "\n";
 
+
+## Loop over the save statements position 
+## 1) Open the file of the structure of the separator 
+## 2) Prepare the temporary mocadi input file to be run  
+## 2 a) Add the prepared header file 
+## 2 b) Open the file of results of the previous mocadi run
+## 2 c) Extract the mean energy of each fragments
+## 2 d) Remove from the filesystem the temporary files of the previous run
+## 2 e) Update the structure of the separator with the new information from the previous mocadi run
+## 3) Run mocadi with the prepared temporary input file of the current iteration
+## 4) Parse the current temporary input file and remove the END statement to save it for the next loop iteration
+## This until the last SAVE statement. The last files will remain as the final updated input file for the mocadi run and the output files of this run.
+
+
 my $previous_save = "#0";
 my $new_header;
 my $previous_out = "0";
-foreach $save (@array_save)
+foreach my $save (@array_save)
 {
     print $save."\n";
-    $namesave=substr $save, 1;
-    $file_in_mocadi = $file_in."_temp_".$namesave.".in";
+    my $namesave=substr $save, 1;
+    my $file_in_mocadi = $file_in."_temp_".$namesave.".in";
 
-    open OUT, ">", $file_in_mocadi or die " impossible d'ouvrir out";
+    open OUT, ">", $file_in_mocadi or die " error to open OUT $file_in_mocadi ! \n";
     print "outfile : $file_in_mocadi \n";
 
-    open IN2, "<", "template_heb.in" or die " impossible d'ouvrir !";
+    open IN2, "<", $template_mocadi or die " error opening $template_mocadi : $! \n";
 
-    %table_fragment=();
+    my %table_fragment=();
 
     if($previous_save =~ "#0")
     {
@@ -84,49 +163,57 @@ foreach $save (@array_save)
 	my $file_out_hbk2 = lc($file_out_hbk);
 	
 	print "process outfile :".$file_out."\n";
-
-	#my $file_in = "C12B10_2GeV_heb.out";
-	print "in file ".$file_out."\n";
-	open IN,"<", $file_out or die "impossible d'ouvrir !";
-	
-	my $frag = 0;
-	my $Energy = 0;
-	my $step = 0;
-	
-	while ($line =<IN>)
+	if($MeanEnergy>0 && $previous_save =~ "#1")
 	{
-	    if($line=~ /expected value.*?([0-9]*) \*\*/)
-	    {
-		$step = $1;
-	    }
-	    if($line=~ /i_fragment.*?([0-9]*)$/)
-	    {
-		$frag = $1;
-	    }
-	    
-	    if(($string_in) = ($line=~ / \<  energy   \>= (.*?) MeV/))
-	    {
-		$string_in =~ /(-?\d.\d+)e(.\d+)/;
-		my ($const, $expon) = ($1, $2);
-		$Energy2 = $const * 10 ** $expon;
-		#print "step #".$step." fragment ".$frag." = ".$Energy."\n";
-		#print $Energy2."\n";
-		$table_fragment{$frag}{$step}=$Energy2;
-	    }
+	    $table_fragment{0}{1}=$MeanEnergy;
 	}
-	close IN;
-
-	unlink $file_out;
-	unlink $previous_out;
-#	unlink $file_out_hbk2;
-
-	for my $k1 ( sort keys %table_fragment ) 
+	else
 	{
+	    print "in file ".$file_out."\n";
+	    open IN,"<", $file_out or die "error opening $file_out: $! \n";
+	    
+	    my $frag = 0;
+	    my $Energy = 0;
+	    my $step = 0;
+	    
+	    while (my $line =<IN>)
+	    {
+		if($line=~ /expected value.*?([0-9]*) \*\*/)
+		{
+		    $step = $1;
+		}
+		if($line=~ /i_fragment.*?([0-9]*)$/)
+		{
+		    $frag = $1;
+		}
+		
+		if(my ($string_in) = ($line=~ / \<  energy   \>= (.*?) MeV/))
+		{
+		    $string_in =~ /(-?\d.\d+)e(.\d+)/;
+		    my ($const, $expon) = ($1, $2);
+		    my $Energy2 = $const * 10 ** $expon;
+		    #print "step #".$step." fragment ".$frag." = ".$Energy."\n";
+		    #print $Energy2."\n";
+		    $table_fragment{$frag}{$step}=$Energy2;
+		}
+	    }
+	    close IN;
+	    
+	    ## this is to remove the temporary file of previous iterration. You can comment them if you want to keep them all
+	    #
+	    unlink $file_out;
+	    unlink $previous_out;
+	    #unlink $file_out_hbk2;
+	    #
+	    
+	    for my $k1 ( sort keys %table_fragment ) 
+	    {
 		for my $k2 ( sort keys %{ $table_fragment{$k1} } ) 
 		{
 		    my $temp_res = $table_fragment{ $k1 }{ $k2 };
 		    print "step #".$k2." fragment ".$k1." = ".$temp_res."\n";
 		}
+	    }
 	}
     }
 
@@ -143,29 +230,24 @@ foreach $save (@array_save)
     my $to_cat = 0;
     my $Energy;
     my $headersave = substr $previous_save, 1;
-    $Energy = $table_fragment{ "1" }{  $headersave };
-    print $Energy."\n";
+    $Energy = $table_fragment{ "1" }{  $headersave }; #"1" normally
+    print $Energy."\n" if(defined $Energy);
     while(my $line2=<IN2>)
     {
-	$line2 =~ s/ TempA / $Af /;
-	$line2 =~ s/ TempZ / $Zf /;
-	if($line2=~/^ SAVE save$previous_save$/)
+	$line2 =~ s/TempA/$Af/;
+	$line2 =~ s/TempZ/$Zf/;
+
+	if($line2=~/^\s*SAVE save$previous_save\s?/)
 	{
 	    $to_cat =1;
 	    next;
 	}
 	if($to_cat==1)
 	{
-	    # if($line2 =~ /^ SAVE save#([0-9]*)$/)
-	    # {
-	    # 	my $headersave = $1;
-	    # 	$Energy = $table_fragment{ "1" }{  $headersave };
-	    # 	print $Energy."\n";
-	    # }
-	    $line2 =~ s/ $Af  , $Zf , 2000./ $Af  , $Zf , $Energy/;
+	    $line2 =~ s/\s*$Af.*?,.*?$Zf.*?,.*?TempEnergy.*?/ $Af , $Zf , $Energy/;
 	    print OUT $line2;
 	}
-	if($line2=~ /^ SAVE save$save$/)
+	if($line2=~ /^\s*SAVE save$save\s?/)
 	{
 	    last; 
 	}
@@ -176,18 +258,24 @@ foreach $save (@array_save)
 
     $previous_save = $save;
 
-    capture sub 
+    # If IO::CaptureOuput is installed or not
+    if($noCapture == 0)
     {
-     	system "./source3.5/mocadi-35 $file_in_mocadi";
-    } => \$stdout, \$stderr;
+	$stdout = capture_exec($mocadi_exe,$file_in_mocadi);
+    }
+    else
+    {
+	#print "Backup Solution !\n";
+	system "$mocadi_exe $file_in_mocadi >| log_LoopMocadi.log";
+    }
+
 
     $previous_out = $file_in_mocadi;
 
     
-    open IN3, "<", $file_in_mocadi or die " impossible d'ouvrir !";
+    open IN3, "<", $file_in_mocadi or die " error opening $file_in_mocadi : $! \n";
     my $temp_new_header;
-    my $Energy;
-    while($linenewheader=<IN3>)
+    while(my $linenewheader=<IN3>)
     {
 	if($linenewheader !~ /^ END$/)
 	{
@@ -195,8 +283,7 @@ foreach $save (@array_save)
 	}
     }
     $new_header = $temp_new_header;
+    close IN3;
 #print $new_header;
     
 }
-
-
